pyproject.toml
[project]
name = "snippet-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Rostislav Smiglic",email = "33193719+Rostislav62@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "3.12.4"
dependencies = [
    "alembic (>=1.14.1,<2.0.0)",
    "aiosqlite (>=0.20.0,<0.21.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "websockets (>=14.2,<15.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
